Move the c directory to cbits
"cabal check" was failing with:

'extra-source-files: c/lz4.h' is not good relative path: same directory segment: ..
'c-sources: c/lz4.c' is not good relative path: same directory segment: ..

As usual, cabal messages are very cryptic. I still do not understand
what it is trying to say, but the error is gone be renaming the "c"
directory to "cbits".